Rename RefFlow service file to Ref
